+++ src/components/AudioPlayer/RepeatAudioModal/RepeatAudioModal.module.scss
@@ -1,0 +1,3 @@
+.separator {
+  margin: var(--spacing-medium) 0;
+}
